backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.0
boto3>=1.34.0
google-generativeai>=0.5.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
weasyprint>=62.0
markdown2>=2.4.0
httpx>=0.27.0
websockets>=12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
